add post style
