+++ src/components/atoms/DotLine.js
@@ -1,0 +1,7 @@
+import React from "react";
+
+export const DotLine = () => {
+  return (
+    <p className="w-full border-dotted border-b border-gray-300 mt-2 mb-2"></p>
+  );
+};
