+++ src/components/organisms/OnlyAuthActions.js
@@ -1,0 +1,26 @@
+import React from "react";
+import { useNavigate } from "react-router-dom";
+
+import EditIcon from "@mui/icons-material/Edit";
+import DeleteForeverIcon from "@mui/icons-material/DeleteForever";
+
+export const OnlyAuthActions = ({ authUser, author, postId, onClick }) => {
+  const navigate = useNavigate();
+  let content;
+
+  if (authUser.displayName === author) {
+    content = (
+      <>
+        <EditIcon
+          className="text-accent"
+          onClick={() => navigate(`/edit/${postId}`)}
+        />
+        <DeleteForeverIcon className="text-danger" onClick={onClick} />
+      </>
+    );
+  } else {
+    content = <></>;
+  }
+
+  return <>{content}</>;
+};
